Break apart interfaces.go, fixes #340
